requirements.txt
streamlit
google-generativeai
python-dotenv
xhtml2pdf
markdown2